requirements.txt
cx-Oracle
pandas
colorama
tqdm
python-dotenv